requirements.txt
Adafruit-DHT==1.3.4
atomicwrites==1.2.1
attrs==18.2.0
certifi==2018.8.24
chardet==3.0.4
idna==2.7
more-itertools==4.3.0
pluggy==0.7.1
psycopg2==2.7.5
py==1.6.0
pytest==3.8.2
python-decouple==3.1
pytz==2018.5
requests==2.19.1
RPi.GPIO==0.6.3
six==1.11.0
urllib3==1.23
